moved our parser to codemod-utils, created test utils that are to be reused in all of our jscodeshift tests in the monorepo
