+++ packages/codemod-utils/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@codemod-com/codemod-utils",
   "author": "Codemod, Inc.",
+  "imports": {
+    "#*": "./src/*"
+  },
   "version": "1.0.0",
   "description": "Set of utilities for jscodeshift",
   "type": "module",
@@ -30,6 +33,7 @@
   },
   "dependencies": {
     "@types/jscodeshift": "^0.11.11",
+    "@babel/parser": "^7.24.4",
     "jscodeshift": "^0.16.1"
   }
 }
