+++ packages/codemod-utils/src/index.ts
@@ -1,6 +1,8 @@
-export * from "./importDeclaration.js";
-export * from "./class.js";
-export * from "./callExpression.js";
-export * from "./react.js";
-export * from "./global.js";
-export * from "./function.js";
+export * from "./jscodeshift/import-declaration.js";
+export * from "./jscodeshift/class.js";
+export * from "./jscodeshift/call-expression.js";
+export * from "./jscodeshift/react.js";
+export * from "./jscodeshift/global.js";
+export * from "./jscodeshift/function.js";
+export * from "./jscodeshift/tests.js";
+export * from "./jscodeshift/parser.js";
