+++ packages/codemod-utils/test/react.ts
@@ -1,30 +1,13 @@
 import assert from "node:assert/strict";
-import jscodeshift, { type FileInfo, type API } from "jscodeshift";
-
+import type { FileInfo } from "jscodeshift";
 import { describe, it } from "vitest";
-import { getClassComponents, getFunctionComponents } from "../src/index.js";
 
-const buildApi = (parser: string | undefined): API => ({
-  j: parser ? jscodeshift.withParser(parser) : jscodeshift,
-  jscodeshift: parser ? jscodeshift.withParser(parser) : jscodeshift,
-  stats: () => {
-    console.error(
-      "The stats function was called, which is not supported on purpose",
-    );
-  },
-  report: () => {
-    console.error(
-      "The report function was called, which is not supported on purpose",
-    );
-  },
-});
-
-const buildRootCollection = (file: FileInfo, api: API) => {
-  const j = api.jscodeshift;
-  const root = j(file.source);
-
-  return { j, root };
-};
+import {
+  buildApi,
+  buildRootCollection,
+  getClassComponents,
+  getFunctionComponents,
+} from "#index.js";
 
 describe("react utils", async () => {
   describe("getClassComponents", async () => {
@@ -45,7 +28,7 @@ describe("react utils", async () => {
         source: INPUT,
       };
 
-      const { j, root } = buildRootCollection(fileInfo, buildApi("tsx"));
+      const { j, root } = buildRootCollection(fileInfo, buildApi());
 
       const components = getClassComponents(j, root);
 
@@ -70,7 +53,7 @@ describe("react utils", async () => {
         source: INPUT,
       };
 
-      const { j, root } = buildRootCollection(fileInfo, buildApi("tsx"));
+      const { j, root } = buildRootCollection(fileInfo, buildApi());
 
       const components = getFunctionComponents(j, root);
       assert.ok(components?.paths().length === 1);
