+++ packages/codemod-utils/tsconfig.json
@@ -1,9 +1,10 @@
 {
-  "include": ["./src", "./global.d.ts"],
-  "exclude": ["./dist", "./build", "./node_modules"],
+  "include": ["./src"],
+  "exclude": ["./dist", "./node_modules"],
   "extends": "@codemod-com/tsconfig/lib.json",
   "compilerOptions": {
     "baseUrl": ".",
+    "rootDir": "./src",
     "outDir": "./dist"
   }
 }
