+++ packages/runner/src/engines/filemod.ts
@@ -12,6 +12,7 @@ import { unified } from "unified";
 import { filter } from "unist-util-filter";
 import { visit } from "unist-util-visit";
 
+import { defaultJSCodeshiftParser } from "@codemod-com/codemod-utils";
 import {
   type CallbackService,
   type Filemod,
@@ -24,7 +25,6 @@ import {
 } from "@codemod-com/filemod";
 import type { Printer } from "@codemod-com/printer";
 import type { Codemod, FileCommand } from "@codemod-com/utilities";
-import { defaultParser } from "#parsers/jscodeshift.js";
 import { isTheSameData } from "#utils.js";
 import type { CodemodExecutionErrorCallback } from "../schemata/callbacks.js";
 
@@ -150,9 +150,11 @@ export const runFilemod = async (options: {
     unifiedFileSystem,
     () => ({
       jscodeshift: jscodeshift.withParser(
-        engineOptions?.parser ?? defaultParser,
+        engineOptions?.parser ?? defaultJSCodeshiftParser,
       ),
-      j: jscodeshift.withParser(engineOptions?.parser ?? defaultParser),
+      j: jscodeshift.withParser(
+        engineOptions?.parser ?? defaultJSCodeshiftParser,
+      ),
       unified,
       hastToBabelAst,
       tsmorph,
